dan_server/requirements.txt
requests==2.28.1
newspaper3k
google-cloud-storage
py7zr
googletrans
lxml
lxml-html-clean
pytz
bs4
newspaper